bds_b3i: working on bug for D1 and D2 messages

+++ src/utils/matlab/libs/gps_l1_ca_read_telemetry_dump.m
@@ -50,7 +50,7 @@ else
     telemetry.preamble_delay_ms = fread (f, count, 'float64',skip_bytes_each_read-double_size_bytes);
     bytes_shift=bytes_shift+double_size_bytes;
     fseek(f,bytes_shift,'bof'); % move to next interleaved
-    telemetry.prn_delay_ms = fread (f, count, 'float64',skip_bytes_each_read-double_size_bytes);
+    telemetry.prn_delay_ms = fread (f, count, 'uint64',skip_bytes_each_read-double_size_bytes);
     bytes_shift=bytes_shift+double_size_bytes;
     fseek(f,bytes_shift,'bof'); % move to next interleaved
     telemetry.Preamble_symbol_counter = fread (f, count, 'float64',skip_bytes_each_read-double_size_bytes);
